requirements.txt
fastapi==0.95.0
librosa==0.9.2
numpy==1.23.0
Pillow==9.4.0
soundfile==0.11.0
uvicorn==0.15.0
